backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2docx
docx2pdf
Pillow
pandas
openpyxl
ffmpeg-python
pyyaml
markdown
pdfkit
